% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mplus_skeleton}
\alias{mplus_skeleton}
\title{mplus skeleton}
\usage{
mplus_skeleton(data, datafile = "mplus_temp.tab")
}
\arguments{
\item{data}{Data frame to extract variable names}

\item{datafile}{Name of the datafile to save}
}
\description{
mplus skeleton
}
