% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runMplusInputAsJob}
\alias{runMplusInputAsJob}
\title{run inp file in Mplus via plugin as a background job}
\usage{
runMplusInputAsJob(x)
}
\description{
run inp file in Mplus via plugin as a background job
}
